+++ packages/@atomiton/nodes/src/index.ts
@@ -4,14 +4,10 @@
  * Clean module exports for the node system
  */
 
-// Base exports - types and factories
 export * from "./base";
 
-// Node implementations
 export * from "./nodes";
 
-// Composite node functionality
 export * from "./composite";
 
-// Type exports
 export * from "./types";